refactor(emulator2): MuMu 启动诊断只在 Emulator 2.0 后端生效
旧 MumuManager 只留一个返回空串的 _describe_launch_failure 挂点，报错文案不变；
诊断的读取与格式化移到 emulator2/mumu6.py 的 MuMu6Manager 覆盖里，和雷电那条
自愈一样只接在 Emulator 2.0 上。
