+++ CHANGELOG.md
@@ -47,7 +47,7 @@
 
 ### 修复
 
-- MuMu 模拟器启动失败或超时时，报错里附上 MuMu 自己给出的启动错误码、错误信息和实例状态，不再只有一句「启动超时」
+- 模拟器 2.0 的 MuMu 实例启动失败或超时时，报错里附上 MuMu 自己给出的启动错误码、错误信息和实例状态，不再只有一句「启动超时」
 - 修复社区通知重复发送、标题重复及同名 Webhook 漏发签到摘要的问题。
 - MAA专项 修复 MAA 内配置方案与新版配置文件不一致时，脚本设置、自动代理与 MAA 更新任务直接报错中断的问题 by [@1w1w11w1](https://github.com/1w1w11w1)
 - 修复库街区签到失败及凭据兼容问题，分别显示游戏与库洛币的签到结果、奖励和失败原因。
