+++ res/version.json
@@ -16,7 +16,7 @@
                 "性能与清理 编辑页连续改动不再丢失，配置文件损坏时保留副本而不是静默清空；日志页、模拟器页、签到页与工具页不再高频轮询，任务日志改为增量推送，MAA 起步不再卡顿一秒，运行日志不再被访问记录撑大；并移除前端从不使用的 9 个接口（MCP 同名工具一并消失）与约两万行无用代码 by [@qiyinxi](https://github.com/qiyinxi)"
             ],
             "修复": [
-                "MuMu 模拟器启动失败或超时时，报错里附上 MuMu 自己给出的启动错误码、错误信息和实例状态，不再只有一句「启动超时」",
+                "模拟器 2.0 的 MuMu 实例启动失败或超时时，报错里附上 MuMu 自己给出的启动错误码、错误信息和实例状态，不再只有一句「启动超时」",
                 "修复社区通知重复发送、标题重复及同名 Webhook 漏发签到摘要的问题。",
                 "MAA专项 修复 MAA 内配置方案与新版配置文件不一致时，脚本设置、自动代理与 MAA 更新任务直接报错中断的问题 by [@1w1w11w1](https://github.com/1w1w11w1)",
                 "修复库街区签到失败及凭据兼容问题，分别显示游戏与库洛币的签到结果、奖励和失败原因。",
